+++ backend/canisters/group_index/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Removed time limit before frozen groups can be deleted ([#3490](https://github.com/open-ic/open-chat/pull/3490))
+- Reduce a few timer job intervals ([#3515](https://github.com/open-ic/open-chat/pull/3515))
 
 ## [[2.0.665](https://github.com/open-ic/open-chat/releases/tag/v2.0.665-group_index)] - 2023-04-17